Makie integration

We currently have very experimental Makie support through GLMakie. GLMakie mostly works around a Screen{T} object to display a scene, where T is some OpenGL-supporting window. GLMakie sets this to a GLFW.Window, but we've made a custom window type to represent a single Figure to be drawn in ImGui. What we get from GLMakie is a framebuffer with a color image texture attachment, and that's displayed by us as an image.

CImGui.MakieFigure — Function

MakieFigure(id::String, f::GLMakie.Figure; auto_resize_x=true, auto_resize_y=false, tooltip=true)

Display a Makie figure in ImGui. See examples/makie_demo.jl for an example of how to use it. This supports all the interaction features in GLMakie:

Scroll to zoom

Click and drag to rectangle select a region to zoom to

Right click and drag to pan

Shift + {x/y} and scroll to zoom along the X/Y axes

Ctrl + left click to reset the limits

Note that scrolling to zoom will also cause the ImGui window to scroll, which can be annoying. This may be fixed in the future by using some other key combination for scrolling to zoom.

These are the interaction events that are wired up and can be used:

hasfocus

entered_window

mousebutton

mouseposition

Known issues:

Changing tick labels don't trigger the scene to be re-layouted, causing them to be clipped if the labels change width. See examples/makie_demo.jl for an example workaround using Makie.tight_ticklabel_spacing!().

The theming doesn't match the ImGui theme so plots look quite out of place by default.

Mouse events aren't delivered unless the mouse is hovered over the figure, so dragging the mouse from within the figure to somewhere outside the figure will keep the old mouse state. e.g. if you're RMB panning and the mouse goes outside the figure, when it enters the figure again panning will resume even the RMB was released.

Drawing can be a bit janky, occasionally the image will not be drawn for a frame or two and you'll see an empty black square instead.

Warning

This is very experimental, you will almost definitely encounter bugs (and if so please submit an issue/PR). We don't consider this covered under semver yet so there may be breaking changes in minor releases.